refactor navbar to ues async promises, split auth servlets - navbar should be loaded before login-specific links
