+++ portfolio/src/main/java/com/google/sps/data/Constants.java
@@ -10,6 +10,7 @@ public final class Constants {
     // mime types.
     public static final String TYPE_JSON = "application/json";
     public static final String TYPE_HTML = "text/html";
+    public static final String TYPE_TEXT = "text/plain";
 
     // name of the comment form field in global.js.
     public static final String FORM_COMMENT = "user_comment";
